:hammer: :newspaper: Create inference file for prediction functions
To separate the raw CLI logic from prediction, a new inference
file is created. This contains logic for predictions. Test_cli
was only testing (now) inference functions so the file was renamed
accordingly.
